Autocompelte Works!
It works on state and zip of the account info page. This page is under the dropdown you get after you sign on and click on your name.

However the styling is really off and I dont't know why. I do not have the time to spend on figuring it out. However it is capable of autocompleting the state and zip